Rename profileStore to publicStore
