+++ src/__tests__/publicStore.test.js
@@ -1,31 +1,31 @@
 const utils = require('./testUtils')
-const ProfileStore = require('../profileStore')
+const PublicStore = require('../publicStore')
 
 const STORE_NAME = '09ab7cd93f9e.public'
 
 jest.mock('../keyValueStore')
 
-describe('ProfileStore', () => {
-  let profileStore
+describe('PublicStore', () => {
+  let publicStore
   const linkProfile = jest.fn()
 
   beforeAll(async () => {
-    profileStore = new ProfileStore('orbitdb instance', STORE_NAME, linkProfile)
+    publicStore = new PublicStore('orbitdb instance', STORE_NAME, linkProfile)
   })
 
   it('should throw if not synced', async () => {
-    expect(profileStore.all('key', 'value')).rejects.toThrow(/_sync must/)
+    expect(publicStore.all('key', 'value')).rejects.toThrow(/_sync must/)
   })
 
   it('should call linkProfile when set is called', async () => {
-    await profileStore._sync()
-    let ret = await profileStore.set('key1', 'value1')
+    await publicStore._sync()
+    let ret = await publicStore.set('key1', 'value1')
     expect(ret).toEqual(true)
     expect(linkProfile).toHaveBeenCalledTimes(1)
     expect(linkProfile).toHaveBeenCalledWith()
   })
 
   it('should return profile correctly', async () => {
-    expect(await profileStore.all()).toEqual({ key1: 'value1' })
+    expect(await publicStore.all()).toEqual({ key1: 'value1' })
   })
 })
